Added experimental persistent configuration handling
The abstract is that each section of the application registers a
struct with the conf package. The conf package loads the previous
values of the struct from persistent storage, allowing for short
and simple access to configurable options.

Mutating these structs should always go through the conf package
as it keeps the global state up-to-date behind the scenes.